+++ TyMD380tools/applet/src/gfx.h
@@ -82,7 +82,7 @@ void print_time_hook(char log);
 
 void get_RTC_time(char* buffer);
 
-uint32_t rgb16torgb(uint16_t color);
+uint32_t rgb16torgb(uint32_t color);
 
 void swapFGBG();
 void swapBG();
